refactor: split rolling risk-free coverage
Evidence:

- python -m pytest tests\unit\test_rolling_mode_adapter.py tests\integration\test_rolling_metrics_endpoint.py -q: 28 passed

- python -m ruff check src\app\services\rolling_risk_free_dependency.py src\app\services\rolling_risk_free_coverage.py scripts\generate_quality_baseline.py: passed

- python -m mypy --config-file mypy.ini src\app\services\rolling_risk_free_dependency.py src\app\services\rolling_risk_free_coverage.py src\app\services\rolling_stateful_inputs.py: passed

- python -m radon mi/cc src\app\services\rolling_risk_free_dependency.py src\app\services\rolling_risk_free_coverage.py: MI A, no C-or-worse complexity

- python scripts\generate_quality_baseline.py: passed and refreshed baseline evidence

- make check: passed, including 551 unit tests
